pyproject.toml
[project]
name = "regexmatch"
version = "0.0.0"
description = "Validates that a value is lower case."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "guardrails-ai>=0.3.2",
    "pydantic>=2.4.2",
    "pytest"
]